Feat/percolator

+++ openaleph_search/index/admin.py
@@ -1,10 +1,7 @@
 from anystore.logging import get_logger
 
 from openaleph_search.core import get_es
-from openaleph_search.index.indexes import (
-    configure_entities,
-    entities_read_index,
-)
+from openaleph_search.index.indexes import configure_entities, entities_read_index
 
 log = get_logger(__name__)
 
@@ -17,7 +14,8 @@ def upgrade_search():
 def delete_index():
     es = get_es()
     log.warning("🔥 Deleting all indices 🔥")
-    for index in entities_read_index().split(","):
+    indexes = entities_read_index().split(",")
+    for index in indexes:
         if es.indices.exists(index=index):
             es.indices.delete(index=index)
 
@@ -25,10 +23,13 @@ def delete_index():
 def clear_index():
     es = get_es()
     log.warning("🔥 Deleting all data 🔥")
-    es.delete_by_query(
-        index=entities_read_index(),
-        body={"query": {"match_all": {}}},
-        refresh=True,
-        wait_for_completion=True,
-        conflicts="proceed",
-    )
+    indexes = [entities_read_index()]
+    for index in indexes:
+        if es.indices.exists(index=index):
+            es.delete_by_query(
+                index=index,
+                body={"query": {"match_all": {}}},
+                refresh=True,
+                wait_for_completion=True,
+                conflicts="proceed",
+            )
